# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "",
        "version": "0.0.0"
    },
    "paths": {
...
        "/api/currency/rates": {
            "get": {
                "tags": [
                    "Currency"
                ],
                "operationId": "get_currency_rates_get_all",
                "parameters": [
                    {
                        "name": "base",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "USD"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Currency Rates List",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CurrencyRateResponseDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/currency/convert": {
            "get": {
                "tags": [
                    "Currency"
                ],
                "operationId": "get_currency_convert",
                "parameters": [
                    {
                        "name": "amount",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 100
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "USD"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "EUR"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Currencies Convertation Results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CurrencyConvertResponseDto"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CurrencyRateResponseDto": {
                "required": [
                    "code",
                    "rate"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "example": "USD"
                    },
                    "rate": {
                        "type": "number",
                        "format": "float",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "CurrencyConvertResponseDto": {
                "required": [
                    "amount",
                    "currency_from",
                    "currency_to"
                ],
                "properties": {
                    "amount": {
                        "type": "number",
                        "format": "float",
                        "example": 92.45
                    },
                    "currency_from": {
                        "$ref": "#/components/schemas/CurrencyRateResponseDto"
                    },
                    "currency_to": {
                        "$ref": "#/components/schemas/CurrencyRateResponseDto"
                    }
                },
                "type": "object"
            }
        }
    },
...
}